+++ src/main/java/utils/Calculatable.java
@@ -9,7 +9,7 @@ import java.util.Locale;
 public interface Calculatable {
     default void writeToCSV(double initialX, double finalX, double eps, double step, String filePath) {
         String csvString;
-        BigDecimal res;
+        double res;
         try(PrintStream printStream = new PrintStream(new FileOutputStream(filePath))) {
             for (double i = initialX; i < finalX; i += step) {
                 res = calculate(i, eps);
@@ -21,5 +21,5 @@ public interface Calculatable {
         }
     }
 
-    BigDecimal calculate(double i, double eps);
+    double calculate(double i, double eps);
 }
